Finishes monsters - implements idleStrategy 'heroLastPosition', difficulty scales with level progress, amount of monsters to spawn scales with level

+++ game/src/ecs/entities/Monster.java
@@ -11,11 +11,12 @@ public class Monster extends Entity {
 
     private float xSpeed = 0;
     private float ySpeed = 0;
-    private float dmg = 0;
+    private int dmg = 0;
 
     /** Entity with Components */
     public Monster(){
         super();
+
         new PositionComponent(this);
        // setupHealthComponent();
     }
